+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file, and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.1 - 2021-12-01
+
+* Fixed a bug that was causing floats to be added to the memory footprint prematurely per resulting in [#125](https://github.com/microsoft/pybryt/issues/125)
+* Refactored internal representation of memory footprints to create abstraction barriers
+
 ## 0.3.0 - 2021-11-17
 
 * Fixed bug in `Value`s tracking `set`s of numbers